allow set vae amp
